+++ Mailer/app/Api/v1/Handlers/Mailers/MailGeneratorHandler.php
@@ -7,7 +7,6 @@ use Remp\MailerModule\Models\Generators\GeneratorFactory;
 use Remp\MailerModule\Models\Generators\ProcessException;
 use Remp\MailerModule\Repositories\SourceTemplatesRepository;
 use Tomaj\NetteApi\Handlers\BaseHandler;
-use Tomaj\NetteApi\Params\InputParam;
 use Tomaj\NetteApi\Params\PostInputParam;
 use Tomaj\NetteApi\Response\JsonApiResponse;
 use Tomaj\NetteApi\Response\ResponseInterface;
@@ -28,16 +27,34 @@ class MailGeneratorHandler extends BaseHandler
     public function params(): array
     {
         return [
-            (new PostInputParam('source_template_id'))->setRequired()
+            (new PostInputParam('source_template_id')),
+            (new PostInputParam('source_template_code')),
         ];
     }
 
     public function handle(array $params): ResponseInterface
     {
         $generator = null;
-        $template = $this->sourceTemplatesRepository->find($params['source_template_id']);
+
+        if (!isset($params['source_template_id']) && !isset($params['source_template_code'])) {
+            return new JsonApiResponse(400, [
+                'status' => 'error',
+                'message' => 'You have to specify either \'source_template_id\' or \'source_template_code\' param to identify generators source template',
+            ]);
+        }
+
+        $template = null;
+        if (isset($params['source_template_id'])) {
+            $template = $this->sourceTemplatesRepository->find($params['source_template_id']);
+        }
+        if (!$template && isset($params['source_template_code'])) {
+            $template = $this->sourceTemplatesRepository->getByCode($params['source_template_code']);
+        }
         if (!$template) {
-            return new JsonApiResponse(400, ['status' => 'error', 'message' => "No source template associated with id {$params['source_template_id']}"]);
+            return new JsonApiResponse(404, [
+                'status' => 'error',
+                'message' => "Cannot find the source template: " . ($params['source_template_id'] ?? $params['source_template_code']),
+            ]);
         }
 
         try {
@@ -53,7 +70,9 @@ class MailGeneratorHandler extends BaseHandler
         }
 
         try {
-            $output = $generator->process($paramsProcessor->getValues());
+            $generatorParams = $paramsProcessor->getValues();
+            $generatorParams['source_template_id'] = $template->id;
+            $output = $generator->process($generatorParams);
             return new JsonApiResponse(200, ['status' => 'ok', 'data' => $output]);
         } catch (ProcessException $exception) {
             return new JsonApiResponse(400, [