+++ Mailer/app/Models/Generators/DennikNBestPerformingArticlesGenerator.php
@@ -60,7 +60,6 @@ class DennikNBestPerformingArticlesGenerator implements IGenerator
     public function apiParams(): array
     {
         return [
-            (new PostInputParam('source_template_id'))->setRequired(),
             (new PostInputParam('articles'))->setRequired(),
         ];
     }