+++ Mailer/app/Models/Generators/MinutaAlertGenerator.php
@@ -46,7 +46,6 @@ class MinutaAlertGenerator implements IGenerator
     public function apiParams(): array
     {
         return [
-            (new PostInputParam('source_template_id'))->setRequired(),
             (new PostInputParam('post'))->setRequired(),
         ];
     }