+++ Mailer/app/Models/Generators/NovydenikNewsfilterGenerator.php
@@ -47,7 +47,6 @@ class NovydenikNewsfilterGenerator implements IGenerator
     public function apiParams(): array
     {
         return [
-            (new PostInputParam('source_template_id'))->setRequired(),
             (new PostInputParam('newsfilter_html'))->setRequired(),
             (new PostInputParam('title'))->setRequired(),
             (new PostInputParam('editor'))->setRequired(),
@@ -256,7 +255,7 @@ class NovydenikNewsfilterGenerator implements IGenerator
         $output = new ArrayHash();
 
         if (!isset($data->post_authors[0]->display_name)) {
-            throw new PreprocessException("WP json object does not contain required attribute 'display_name' of first post author");
+            throw new PreprocessException("WP json object does not contain required attribute 'post_authors.0.display_name'");
         }
         $output->editor = $data->post_authors[0]->display_name;